Improve test_valid_complex_string_with_space() xfail condition
Improve the xfail condition for `test_valid_complex_string_with_space()`
to use PyPy version rather than Python version.  This corresponds
to the fact that the issue was fixed in 7.3.17 releases.  The other
condition works only because PyPy3.9 is no longer being developed,
and therefore did not get a 7.3.17 release.