Allow collections to be passed instead of just sets for `include`/`exclude`
The idea is to try to first convert the arguments to a set if provided,
not a dict, and iterable. If not, it is returned as is. It turns out
that later in the code, we also allow containers to be used (although
not documented and not allowed in the `IncEx` type alias). This explains
why until now, we could still pass lists/tuples for these arguments,
but it would not behave as expected, especially in `map_negative_indices`,
only checking for sets.

With the proposed changes, include and exclude are guaranteed to be either
None, a dict, a set, or a container (implementing `__contains__`).